LUS-10721 mdd: fix parent FID in changelog of striped directory
Changelog entry for the file operations such as create, rename,
link, unlink, mkdir referred to parent FID ("p=") as a shard's
FID in a striped directory. The same was true for the source's
parent FID ("sp="). This commit hides the Lustre intrinsics from
user displaying the parent's directory FID instead as expected.

Obtaining the parent FID via the linkEA from a remote object is an
expensive operation, so this functionality needs to be explicitly
enabled via a tunable and is disabled by default.

Signed-off-by: Dmitry Ivanov <dmitry.ivanov2@hpe.com>
Change-Id: Ic627795d6df02b52c1f4c8e4ca1bfcac4e007f49
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/160654
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Shaun Tancheff <stancheff@cray.com>
Reviewed-by: Vitaly Fertman <vitaly.fertman@hpe.com>