LUS-11090 lod: stop using EA with max_inherit==0
stop using EA with max_inherit==0 as it has no sense. if the feature
is to be disabled for a particular subdir, as lfs-setdirstripe(1) says
the following can be used:
	lfs setdirstripe -D -i 0 -c 1 --max-inherit 1 <dir>

it also prevent problems like inheritance in lod_ah_init() from a
parent with a default with max_inherit==0, what is not checked, thus
inherited; and similar problems on client.

Signed-off-by: Vitaly Fertman <c17818@cray.com>
Change-Id: I9225701637c002c2c9c6fa1491b34f87d9ddf5f8
Reviewed-on: https://es-gerrit.dev.cray.com/161048
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>