LUS-11415 kfilnd: Wrong traffic class assigned
The wrong traffic class was being assigned to a transmit
context when multiple networks were assigned to the same
interface.

This was discovered by noticing a currently unsupported
traffic class didn't fail when it was used. The traffic
class from the shared domain was being used instead.

This was fixed by explicitly specifying the traffic
class when creating a transmit context for an endpoint.

Test-Parameters: trivial
Signed-off-by: Ron Gredvig <ron.gredvig@hpe.com>
Change-Id: I21236c01d4bef53b62e2f303c8e24e059ce83c0a
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/161664
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Ian Ziemba <ian.ziemba@hpe.com>
Tested-by: Chris Horn <hornc@cray.com>