LUS-11330 test: check the inherited pool for conflicts
in addition to LU-15658, the start index could be inherited from
parent and the pool from root: drop the pool in case of conflict
as well.

Also, let the stripe_count to be adjusted by the count of targets
in the pool only if the pool is not dropped.

Another case of a problem inheritance is saving the inherited LOVEA
to subdir, when all the parameters are inherited but the ost list.

Signed-off-by: Vitaly Fertman <vitaly.fertman@hpe.com>
Change-Id: I34b8a4fbf0d412d17d922369e818ad218387e57a
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/161799
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>