LUS-11505 kfilnd: kfilnd_peer ref leak on send
There is an extra refcount_inc() done by kfilnd_tn_alloc_for_peer().
This is correct in the case where we are allocating TN for HELLO
request, because our caller does not take extra ref on kfilnd_peer,
but it is wrong in the normal kfilnd_tn_alloc() path because
kfilnd_tn_alloc() takes this reference by way of a call to
kfilnd_peer_get().

Move the refcount_inc() from kfilnd_tn_alloc_for_peer() to
kfilnd_send_hello_request() where it is needed.

Test-Parameters: trivial
Fixes: bad47a5f36 ("LUS-11128 kfilnd: Allow one HELLO in-flight per peer")
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I3d723a829ec42929ce22a80ffda97dbd87917d4b
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/161860
Reviewed-by: Ron Gredvig <rgredvig@cray.com>
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>