LUS-11569 hsm: yield cdt_lock while updating actions llog records
The CDT thread holds cdt_lock in write mode while scheduling requests or
updating records in the HSM actions llog; if there is a large number of
requests, other MDT threads that also need to take the lock in order to
e.g. mark an action as complete, add a new action, or print the contents
of the llog can be starved and get blocked. Yield cdt_lock whenever a
certain number of requests have been handled to avoid this. The value
for the number of requests is based on field reports of the issue and
some experimentation; it may not be optimal but should hopefully help
address the issue.

Change-Id: Id607e1ac9720a954ecda1fb38b4b22ef491c4549
Signed-off-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162356
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>