LUS-11572 kfilnd: Handle TAG_RX_OK in TN_STATE_FAIL
It is possible for the fabric to delay packets such that the retry
handler cancels the message but it is still delivered to the target.
If the timing is right then the initiator may receive a TAG_RX_OK
event after the transaction has transitioned to TN_STATE_FAIL. This
currently trips an LBUG, but instead we can allow the transaction to
complete normally.

Test-Parameters: trivial testlist=sanitiy-lnet/302
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I381d64713a7942fed09d41b30f64be602193057f
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162358
Reviewed-by: Ian Ziemba <ian.ziemba@hpe.com>
Reviewed-by: Ron Gredvig <rgredvig@cray.com>
Tested-by: Vitaly Fertman <c17818@cray.com>