LUS-11265 lod: improve rr allocation
Roundrobin allocation uses atomic_inc() % ost_count for
generation OST index. When some OSTs are unavailable and
many threads make object creation, it could happen that
OST idx is the same for all attempts. For example with
4 OSTs configuration when 2 OSTs do faiover, estimation
of probability is 0.5^12=0.024%. The result is ENOSPC for
user application.

Let's try one by one OSTs for a last speed loop.

Signed-off-by: Alexander Boyko <alexander.boyko@hpe.com>
Change-Id: If960f23780548dd652947db77d54f23d3c304d21
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162362
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>