LUS-11632 libcfs: cfs_num_min_max() returns wrong max value
The cfs_num_min_max() function is called using the nf_min_max function
pointer. It returns the minimum and maximum values for a given
range. However, if a network type uses cfs_num_min_max()
to find the min and max of the range, it will encounter an issue
where the start and end values of the range will be same.

This occurs because cfs_num_min_max() was returning the same value
for both the max and min

Here's an example specific to the "kfi" type, which utilizes
"cfs_num_min_max()" to find the minimum and maximum values.
As a result of the bug, the start and end values of
the range are the same.
Added range for kfi type--
lctl nodemap_add_range --name alps --range [0-131071]@kfi
show the range added
lctl get_param nodemap.alps.ranges
nodemap.alps.ranges=
[
{ id: 1, start_nid: 0@kfi, end_nid: 0@kfi }
]

Signed-off-by: Rajeev Mishra <rajeevm@hpe.com>
Change-Id: I3a6ccd8e87ff438a160815f2147a43071e82304d
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162394
Reviewed-by: Shaun Tancheff <stancheff@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>