LUS-11641 cray-obs: subproject cray-lustre-client-dkms
Incorporate the dkms build from cray-lustre-client-dkms.spec
into cray-lustre-client.spec directly.

This also updates some Requires: dependencies in
cray-lustre-client.spec and cray-lustre-client-ofed.spec
to work with either the kernel module packages or the dkms
installed package.

Update dkms scripts to pass kmoddir and auto detect
ofa_kernel and kfabric and pass the appropriate options
to configure.

Update installation of lustre and lnet headers.

Remove cray-lustre-client-dkms.spec as it was not used
and is no longer needed.

Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I2e535d0f6ab50f7549ae6e95bbd07420961e7920
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162402
Reviewed-by: Rajeev Mishra <rajeevm@hpe.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>