LUS-11808 ldiskfs: Periodically write ldiskfs superblock
This patch introduces a mechanism to periodically check and update
the superblock within the ext4 file system. The main purpose of this
patch is to keep the disk superblock up to date. The update will be
performed if more than one hour has passed since the last update,
and if more than 16MB of data have been written to disk.

This check and update is performed within the
ext4_journal_commit_callback function, ensuring that the superblock
is written while the disk is active, rather than based on a timer
that may trigger during disk idle periods.

Lustre-change: https://review.whamcloud.com/51340
Lustre-commit: e27a7b33d6351ff8b8bae101079af88f4eedac99
Lustre-author: Vitaliy Kuznetsov <vkuznetsov@ddn.com>

WC-bug-id: LU-16298
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I85896b64d19b16f114c3a51f71b36e9a52230c18
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162510
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>