LUS-11466 tests: file_truncate: command not found part 2
Patch contains the changes appeared during upstreaming
because of Test-Parameters used in Autotest:
(LU-16521 tests: allow + separator for racer)

The Test-Parameters: line parses ',' even with quotes, so it cannot
be used as a separator in Autotest for RACER_PROGS and RACER_EXTRA.

Allow both ',' and '+' as a separator for both RACER_PROGS and
RACER_EXTRA tasks so specific racer tasks can be run.

Do not always enable dir_remote and dir_migrate if RACER_PROGS set.

Test-Parameters: testlist=racer env="DURATION=120"
Fixes: 6d9e74580e ("LU-14274 tests: enhance racer to set extra layout")
Signed-off-by: Elena Gryaznova <elena.gryaznova@hpe.com>
Change-Id: If587dbe07326dead215d34cd1fa8908c1b4d1dbf
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162535
Reviewed-by: Andriy Skulysh <andriy.skulysh@hpe.com>