LUS-11893 build: Conditionally install kkfilnd.ko
The dkms.conf file is responsible for installing the built
modules into the dkms module tree. However this file is created
before configure is run.

Because kkfilnd.ko and kgnilnd.ko modules are not always built
these need to have conditional logic to include them as
modules to be installed.

Unfortunately dkms.conf does not get any useful information
from the configure and build steps so we duplicate the
expected logic passed to configure.

If a reasonable /usr/src/kfabric/[<flavor>|default] exists then
kkfilnd.ko is expected.

Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I328c67fbc27c53dfa605f2587cf2b152bf8d33f8
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162550
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Tested-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>