LUS-10363 build: Conditionally require kmod-zfs-devel
Server with zfs support requires either kmod-zfs-devel
or a configure that points to the required headers and
library files via configure.

Here we check the configure arguments for '--with-zfs-obj='
if the zfs path is specified for configure the package
requirement is not needed.

Otherwise require the kmod-zfs-devel package and require
one of libzfs-devel, libzfs4-devel or libzfs5-devel

Lustre-commit: 77d01c485e00468b2ef6e3eb64544d30c049a411
Lustre-change: https://review.whamcloud.com/46356

Test-Parameters: trivial fstype=zfs
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I2f08d1b97813a77ac02ac598fe6c3305723ae3c4
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/161635
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>