LUS-11965 lov: fix fiemap past EOF failure
cp occasionally fails with EINVAL from fiemap.
Tracing shows that cp requests extents with
offsets past the end of file. However, the
fiemap documentation from the kernel source tree
does not specify that we should return an error
in this case.

The fix is to return 0 mapped extents. Lustre
already does that for zero size files, in
ll_do_fiemap().

Change-Id: I1cc1900ddf393be3c80517da100511027ef7612c
Signed-off-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162595
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>