LUS-11711 build: Add in-kernel-ko2iblnd driver
Add in-kernel-ko2iblnd.ko for users of in-kernel OFED
and only build ko2iblnd.ko if an external OFED is available

This allows for building and packaging both an external
(MOFED or OPA) o2ib driver and an in-kernel o2ib driver.

Packaging rules will be written so that only enable one
of the o2iblnd drivers can be installed.

In the case of the in-kernel-ko2iblnd.ko driver a symlink
named ko2iblnd.ko will be created to point to the in-kernel
based o2ib driver which allows for a reasonable migration path
for the majority of users.

It is useful for dist build and test to be able to build
both in-kernel IB and external OFED in the same build.

This also means there would be some install/configure
adjustments that ought to have some discussion.

Lustre-change: https://review.whamcloud.com/c/51915
Lustre-commit: 8b1d2a72f10e8486ff83e4c5a22420232246b582

WC-bug-id: LU-16967
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I1c86dd38f81d243c07eddab096bc2f865987493f
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162727
Reviewed-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>