LUS-7141 ptlrpc: skip unnecessary client eviction
Backport of version of the fix from LU-14708.

If a client has not cached any data from a server and has not
performed any modifying rpcs to the server - let the client to connect
as a new one instead of considering itself as evicted.

recovery-small:test_150c is removed as contradicting to the fix logic.

Signed-off-by: Vladimir Saveliev <vladimir.saveliev@hpe.com>
Change-Id: Id087d1e5be61f656da374c19d7008c0ac5d98c6b
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162606
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>