LUS-12086 ptlrpc: replay open outside a list
Open requests which pinned by closes have old transaction number
and less than last_commited. It creates a gap in transaction
sequence and leads to a client reply and waiting request from it.
So recovery takes more time.
Open requests could be processed outside of the replay queue,
to skip a gap handling, also it would be parallel processing.

Signed-off-by: Alexander Boyko <alexander.boyko@hpe.com>
Change-Id: I68e61e43e6730c661d6567f2842b470322da714d
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>