LU-17421 build: Update check for arc_prune_func_t parameters
In OpenZFS 2.2.1 the code for arc_prune_async() was unified so that
FreeBSD and Linux did not have their own implementation versions of
the same code. Part of this update changed first parameter for the
arc_prune_func_t to be an uint64_t.

Without this patch, Lustre would not build with ZFS 2.2.1 because of
a failure for incompatible pointer types for the arc_prunte_func_t
function pointer passed to arc_add_prune_callback().

Test-Parameters: trivial fstype=zfs
Signed-off-by: Brian Atkinson <batkinson@lanl.gov>
Change-Id: Iaa03cc9421f27a8517ce04817f04102de9adb86a
Reviewed-by: Akash B <akash-b@hpe.com>
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162722
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Vitaly Fertman <vitaly.fertman@hpe.com>