LUS-11702 hsm: resend pending requests on agent registration
When a copytool process is terminated abruptly while handling HSM
requests, it is not possible for any respawned copytool process to
resume handling the incomplete HSM requests, as they are already marked
with HS_* flags that prevent subsequent HSM requests to be issued upon
the files, unless the HS_* flags are cleared manually.

This patch allows to avoid requiring manual intervention in such cases,
by keeping a record of in-flight HSM requests per-agent, detecting that
the agent had been registered beforehand and resending the incomplete
HSM requests to the copytool at registration time.

Change-Id: I1ee4c3c336c7096f1e373e46fe2c4e86ba9f2266
Signed-off-by: Nikitas Angelinas <nikitas.angelinas@hpe.om>