LUS-12097 dkms: improve /etc/sysconfig/lustre
Expand the features available in /etc/sysconfig/lustre
to enable more flexability to dkms users.

Providing y/n switches for common features:
    LUSTRE_DKMS_ENABLE_GSS=y/n
    LUSTRE_DKMS_ENABLE_GSS_KEYRING=y/n
    LUSTRE_DKMS_ENABLE_CRYPTO=y/n
    LUSTRE_DKMS_ENABLE_IOKIT=y/n

As well as a catch-all to pass to configure:
    LUSTRE_DKMS_CONFIGURE_EXTRA='string passed to configure'

Add suport for dpkg checking for libkrb5-dev to enable or
disable gss by default, if it is not otherwise specifed.

Lustre-commit: [TBD: 24ead1ca7418bd86736851439c266a915a8c3941]
Lustre-change: https://review.whamcloud.com/c/53787

Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I526ccf264450121bb62033ba52e300d0047e0f3a
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162730
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Reviewed-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>