LUS-11793 utils: setstripe overstripe multiple OST count
Add an option to "lfs setstripe -C" to specify stripe counts
that are a multiple of the number of OSTs in the filesystem.
Using "-C -1" will create one stripe on all (available) OSTs,
as with "-c -1", to avoid too many stripes.  Using "-C -2"
will create two stripes on each OST, etc.

The maximum multiplier is currently "-C -32", which will
create 32 stripes per OST. It is still possible to specify
a large positive stripe count directly to  "-C" for testing
purposes and to maintain compatibility with current usage.

Signed-off-by: Rajeev Mishra <rajeevm@hpe.com>
Change-Id: Ibc27ed6cbb8743922b6c5785f65b86974e322773
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162937
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>