LUS-11742 tests: explicit check for eviction with dmesg parse
client_evicted() used to check for client eviction based on result of
lfs df. When it returned any error but EOPNOTSUPP - that was taken as
"client was evicted".

When glibc's realpath() changed to not call stat()
(see for ref
stdlib: Sync canonicalize with gnulib [BZ #10635] [BZ #26592] [BZ
..
- Realpath mishandles EOVERFLOW; stat not needed anyway (BZ#24970).
)
lfs df started to return EOPNOTSUPP from lfs_df(). Than
client_evicted() was changed (LU-16916 tests: fix client_evicted() not
to ignore EOPNOTSUPP) and now any not 0 of lfs df is taken as client
was evicted.

Check for "This client was evicted" in dmesg output to make sure that
eviction happened.

Signed-off-by: Vladimir Saveliev <vladimir.saveliev@hpe.com>
Change-Id: Ib5e849a874f8136f01aba39e44dae8d94134cc59
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162703
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Elena Gryaznova <elena.gryaznova@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>