LUS-12451 ldiskfs: restore ldiskfs patch attribution
Over time various ports of ldiskfs patches to newer kernel
releases and distribution kernels have lost or confused the
original history and author of many patches.  It is also
helpful to have a summary for the reason behind each patch.

Thanks-to: Andreas Dilger <adilger@whamcloud.com>
for digging through the history of these patches.

Lustre-change: https://review.whamcloud.com/54247
Lustre-commit: 0777144705f0abf351dedb5262aea9aa71903d38

HPE-bug-id: LU-17599
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I7154f9e3b890e7fa2c029c400af48caf2101b6fc
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163013
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>