LUS-12108 llapi: restore 'pool=ignore' functionality
Changes to llapi_stripe_param_verify() and related llapi file
creation functions to verify that the given pool name is valid
introduced a bug that disallowed the 'ignore' pool name, which
is used to create files without any pool name.

Allow the reserved pool names from lov_pool_is_reserved() to be
used even (especially!) if the named pool does not exist.

Revert the changes to ost-pools.sh::test_32() that created the
'ignore_pool' pool, and go back to checking that 'ignore' will
create a file that does not use any pool.

Change the pool name validation to only do fsname lookup if the
pool name is actually specified, instead of looking up fsname
but not actually using it for anything.

Lustre-change: https://review.whamcloud.com/54355
Lustre-commit: 61c232554dae25340369e19baa98424e76288a22

Fixes: 83c2c486f8 ("LUS-11501 llapi: Verify stripe pool name")
Signed-off-by: Rajeev Mishra <rajeevm@hpe.com>
Change-Id: Ic1f263708addb89a153040693032c94f47719e0a
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162908
Reviewed-by: Vitaly Fertman <vitaly.fertman@hpe.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>