LUS-11686 llite: get rid of ci_dont_repeat flags
This is to sync up with version of the fix landed into master.

@ll_page_inv_lock is used to check the page being read by kernel
after it has been deleted from Lustre, which avoids potential
stale data reads. This seqlock allows us to see that a page was
potentially deleted, catch it in this case and repeat the I/O in
ll_filemap_fault() or vvp_io_read_start().

To avoid endless I/O repeat wrongly, in this patch we only increse
@ll_page_inv_lock for the page in PageUptodate state when delete
the page in vvp_page_delete(). The page that not in PageUptodate
state is usually deleted due to the error that does not require
retry.
By this way, ll_filemap_fault() and vvp_io_read_start() will not loop
endless for those errors that does not need to repeat I/O as the
seqlock @ll_page_inv_lock does not have any change.

Signed-off-by: Vladimir Saveliev <vladimir.saveliev@hpe.com>
Change-Id: I2345e0e8d24af8a492d3e631fc71a490f6bbd834
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162837
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>