LUS-10060 build: always include git hash in commit
In some cases, git describe may omit the git hash
if an exact match of a tag is found.
For example, when on tag 2.15.3.3_cray,
'git describe --match "[0-9]*" --abbrev=7 HEAD' produces
'2.15.3.3_cray', when the commit hash 12cc2d7 on the end is
desired: '2.15.3.3_cray-0-g12cc2d7'.

This adds the --long option to git describe to include the
commit hash on the version description (shortened to 7 chars).

Signed-off-by: Caleb Carlson <caleb.carlson@hpe.com>
Test-Parameters: trivial
Change-Id: Ica6ac0d9c29b9bde4a82eb0973c2de095815603a
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162912
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Sakib Samar <sakib.samar@hpe.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>