LUS-12356 build: detect mlnx-ofa_kernel-devel contents
Parse the configure_args for --with-o2ib and allow the
user specified path to override the mofed defaults.

Further when mlnx-ofa_kernel-devel contents are available silence
the BuildRequires: to allow for an mlnx source installation to
satisfy the lustre build requirements.

In addition move the mlnx specfic requirements to the mofed lnd
when '--with multiple_lnds' is enabled.

Lustre-commit: 3c66185c84c94c47f5d8031b92165c68ef666e0e
Lustre-change: https://review.whamcloud.com/53613

Fixes: 67cd54d05d ("LU-16967 build: Separate lnet LND rpm packaging")
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Ic21425bcd3384d3e57377353af93fe1d6ebd6121
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162914
Reviewed-by: Caleb Carlson <caleb.carlson@hpe.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>