LUS-12378 build: restore older RPM support
RPM versions prior to 4.13. do not include 'or' keyword
support for Requires.

For such clients (RHEL 7, SUSE 12) disable the
'lustre-client-dkms or kmp' leaving only the kmp as was
previously the case.

Also fallback for ZFS one of:
  --define 'zfs 5'
  --define 'zfs 4'
can be used to target ZFS >= 2.1 or 2.0 respectively,
leaving the default to be ZFS 0.8.

Also loosens user space requirement to require a Lustre version
>= major.minor

Fixes: c65db7482a3 ("LUS-12246 lnet: Add lnet.pc")
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Ib26685bcb8335f7dc36312894a26701311610148
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/162933
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>