LUS-12927 ldiskfs: add NO_CHECKS flags for ext4_iget
Since v6.4-rc4-1-gb3e6bcb94590, the upstream kernel introduced
a strict checking of EXT4_IGET_EA_INODE and it has to match the
EXT4_EA_INODE_FL from the inode.

This creates a problem for osd scrub as we iterate over the inodes
according to bitmap we have no idea if the inode is an ea_inode or
not.

Add a new flag EXT4_IGET_NO_CHECKS and relax the strict checking
when the flag is used.

Cleanup osd_ldiskfs_iget_special(), the caller of osd_iget() now are
passing in the additional flags to ext4_iget().

Lustre-author: Li Dongyang <dongyangli@ddn.com>
Lustre-change: https://review.whamcloud.com/55896
Lustre-commit: a7239d80fc92380a5a56ff1b0adb99d349f9a0dc

WC-bug-id: LU-18080
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: If1a8ea59d9a883e36179df60f56764a70932ec0e
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163357
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>