LUS-12467 build: fix 'or' clause for rpmbuild
This fixes the 'or' clause testing for older distro
versions, which was intended to provide backwards
compatibility for builds on distros that didn't natively
support the rpmbuild 'or' keyword. Instead, it always
evaluated to true, making it so that the DKMS package
was never included as a requirement satisfaction.

This change uses the 'or dkms' Requires, if any of the
distro versions are above a certain level.

Fixes: 5b062866922 ("LUS-12378 build: restore older RPM support")
Signed-off-by: Caleb Carlson <caleb.carlson@hpe.com>
Test-Parameters: trivial
Change-Id: If9666cc45b34f83f5b6cc0b0ebea6bef68af99c1
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163021
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>