LUS-12499 osd-ldiskfs: pass through resize ioctl
Pass through the EXT4_IOC_RESIZE_FS ioctl to the underlying ldiskfs
code so that it is possible to online resize MDT and OST filesystems.

When running resize2fs against a filesystem, it compares st_rdev of
the block device against st_dev of the mounted filesystem, so the
mounted Lustre stub filesystem needs to return proper stat information
from the ldiskfs root directory.  Add in a server_getattr() method
to the server inode_operations.  Using generic_fillattr() is enough,
we don't need the added complexity of calling ext4_getattr() (which
does not exist on directories for all kernel versions).

Change the OSD API from returning the superblock with dt_mnt_sb_get()
to returning the vfsmount with dt_mnt_get(), since it also contains
the superblock, but is more useful for calling some inode methods.

Lustre-change: https://review.whamcloud.com/20161
Lustre-commit: ac0380dc519aa15310670d164e98453861ef332a
Lustre-author: Andreas Dilger <adilger@whamcloud.com>

WC-bug-id: LU-137
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Ib869696d068f72f7de7d99b4ac44118b09e23406
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163043/
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>
Tested-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>