LUS-12515 tests: a test case for write vs truncate deadlock
The patch add a new test case to sanityn.sh which reproduces
the write vs truncate deadlock when group locks are in use.

The test case writes, under the group lock, to a file, pauses
the BRW processing on the OSS side after the page lock is
taken but before ofd_read_lock/osd_trunc_lock is taken.
In parallel, a truncate is sent under a group lock, truncate
get the ofd_write_lock and attempts to lock the same page.
BRW unpauses, attempts to get ofd_read_lock/osd_trunc_lock
and the deadlock is finally reached.

The deadlock is possible with either
[lock_page->ofd_read_lock vs ofd_write_lock->lock_page]
or
[lock_page->osd_trunc_lock(SHR) vs osd_trunc_lock(EXC)->lock_page]

The test reproduces the first one since it's easier.

Change-Id: I9553075a93ded68219346bdc816c98072ffc0706
Signed-off-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163070
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>