LUS-12539 lnet: Use only NI status for route aliveness
lnet_is_route_alive() considers a route to be alive if the associated
gateway has at least one peer NI on both the local and remote nets
that is alive. A peer NI is alive if the NI status is UP and its
health value is at LNET_MAX_HEALTH_VALUE.

Unfortunately, LNet health values do not accurately reflect whether a
route is actually up or down. For example, when a router is forwarding
a PUT w/ACK to a peer that is down, the ACK will not be sent back to
the message originator. This will result in the originator
decrementing the health value of the router peer NI. If this is the
only NI on the local net then this route will be considered dead until
the health value recovers.

It is typical for the number of clients to greatly outnumber the
number of routers. When many clients are doing IO, and those clients
crash or otherwise become unresponsive, this can result in servers
temporarily losing many routes as they hit response timeouts.

To address this issue we no longer consider health values when
determining whether a router peer NI, and thus a route, is considered
alive. Instead we rely solely on the NI status.

Health values of router peer NIs are now considered in
lnet_compare_gw_lpnis() so that we will prefer healthier routes.

lnet_find_route_locked() is lightly re-factored to remove some
code duplication.

The description of router_sensitivity_percentage is modified to note
that this parameter is now deprecated.

Test-Parameters: trivial testlist=sanity-lnet
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Ic27b440bd296686262d4f3bad80cd0de95a32540
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163100
Reviewed-by: Caleb Carlson <caleb.carlson@hpe.com>
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>