LUS-12556 lnet: Graceful router removal and addition
Removing an LNet router from service (say, for repairs) on a
production system is somewhat difficult today.

Current best practices involve either removing all routes for a router
from all clients and servers, then waiting for the router
buffers to drain before shutting down LNet, or disabling routing on
the router node itself. The former is not ideal because a route may be
removed on one side or another at different times creating an
asymmetric route, and this can result in network failures.
The latter is not ideal because the router will immediately drop any
buffered messages as well as new messages that arrive.

This patch provides a better solution for graceful shutdown by
modifying the behavior of LNet when routing is disabled via lnetctl.

A new routing state, LNET_ROUTING_STOPPING, is defined to indicate
that the router is in the process of shutting down. While in this
state, the router forwards messages as usual, but the
LNET_PING_FEAT_RTE_DISABLED bit is set in the router's ping buffer
feature bits. When clients and servers discover the router they will
see this bit and stop using the router for future sends.

To provide time for the clients and servers to stop using the router,
the monitor thread takes over responsibility for freeing the router
buffer pools. When the thread sees that the router has entered the new
STOPPING state, and after alive_router_check_interval +
router_ping_timeout seconds have passed, it checks whether any of the
router buffer pools are still in-use. If none are in use then it will
free the buffers and transition the routing state to
LNET_ROUTING_DISABLED.

If the pools are still active after 3x alive_router_check_interval +
router_ping_timeout seconds, then the monitor will drop any pending
messages, free the buffers and transition the routing state to
DISABLED.

Similar issues may occur when bringing a router back into service. If
the routes were removed then they must be added back. This can result
in asymmetric routes. It has also been observed that a router can
struggle to handle the influx of connection requests when it is
initially returned to service. This can result in failure to forward
messages when a connection cannot be established to the message
target.

Our solution is similar as above. A new LNET_ROUTING_STARTING state is
defined to indicate the router is in the process of starting up. The
monitor thread takes over responsibility for transitioning to the
LNET_ROUTING_ENABLED state and clearing the
LNET_PING_FEAT_RTE_DISABLED bit. The monitor thread will wait for
alive_router_check_interval + router_ping_timeout seconds before
enabling routing. This provides time for peers to establish a
connection with the router before they all start using it as a
next-hop.

Test-Parameters: trivial testlist=sanity-lnet
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Idcd2c68028f60b760f6cc503291aa54f78042ca1
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163105
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Caleb Carlson <caleb.carlson@hpe.com>