LUS-12253 osd-ldiskfs: use local fake file for iteration
LU-13783 introduced a alloc_file_pseudo to directory iteration.
It leads to many file descriptors for a kernel and a special
logic to drop it LU-16973.
"LU-16973 ptlrpc: flush delayed file desc if idle" cause a
mdtest directory remove performance degradation of up to 28%.

One of the reason of alloc_file_pseudo was a security_alloc() call.
However we could initialize iteration like kernel and skip
security checks for a pseudo files.

Fixes: 9e92a12013 ("LU-16973 ptlrpc: flush delayed file desc if idle")
Fixes: b0f150eba4 ("LU-13783 osd-ldiskfs: use alloc_file_pseudo to create fake files")
Signed-off-by: Alexander Boyko <alexander.boyko@hpe.com>
Change-Id: I255c6d8ace6de00dd94a704c800a6076f35f3da5
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/#/c/162892/
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>